requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
jinja2==3.1.3
openpyxl==3.1.2
reportlab==4.0.8
httpx==0.26.0
alembic==1.13.1
python-magic==0.4.27
alembic
bcrypt
email-validator
fastapi
gunicorn
httpx
jinja2
openpyxl
passlib[bcrypt]
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
python-magic
python-multipart
reportlab
sqlalchemy
uvicorn[standard]
alembic
bcrypt
email-validator
fastapi
gunicorn
httpx
jinja2
openpyxl
passlib
psycopg2-binary
pydantic
pydantic-settings
python-jose
python-magic
python-multipart
reportlab
sqlalchemy
uvicorn
alembic==1.13.1
bcrypt==4.0.1
email-validator==2.1.0
fastapi==0.109.0
gunicorn==21.2.0
httpx==0.26.0
jinja2==3.1.3
openpyxl==3.1.2
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
reportlab==4.0.8
sqlalchemy==2.0.25
starlette
uvicorn[standard]==0.27.0
alembic==1.13.1
bcrypt==4.0.1
email-validator==2.1.0
fastapi==0.109.0
gunicorn==21.2.0
httpx==0.26.0
jinja2==3.1.3
openpyxl==3.1.2
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
reportlab==4.0.8
sqlalchemy==2.0.25
starlette
uvicorn[standard]==0.27.0
alembic
bcrypt
email-validator
fastapi
gunicorn
httpx
jinja2
openpyxl
passlib[bcrypt]
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
python-magic
python-multipart
reportlab
sqlalchemy
starlette
uvicorn[standard]
